Update Permissions: Add missing permissions cause
The added cause was found on the discord server during support and not a thing we previously thought about.
Somehow the prefixing of a decimal with the hex indicator `0x` causes the API to reject the patch request for the overwrite update.  
  
Note: I'm not too familiar with the internal workings at this point and can just assume that the bots permissions are checked against the invalid permission number resulting in a rejected request. Since invalid permission values are usually accepted and used as-is and as applicable I'm not entirely sure of the underlying circumstances causing an error instead.