Note to not use the Client Secret
Add a note to not use the Client Secret from the General Information panel.

In the event the user ignores the warning to read carefully (mea culpa) and is used to apps only having one secret token, it's possible to confuse the Token with the Client Secret.  If you end up in that spot, you'll get an invalid token response from Discord and be very confused.