refactor: adjusted wording to be more concise
This refactor adjusts some of the wording to ensure a clean and concise definition of what either "ping" approach they take, along with fixing the text in between the <!-- eslint-skip --> and the code block. There is also a tip added for the v12 branch that provides information on retrieving a certain shards heartbeat. The brief explanation for the second example has been adjusted to provide an explanation on what is happening.