Update v11 -> v12 changes for client#shardResume
In the guide, it states that client#shardResume has the parameters "replayed" and "shardID". These are in fact backwards, as in the docs it states the parameters are "shardID" THEN "replayed". Tested this with my own bot to confirm that the order was incorrect for the guide instead of the docs.