Fix: Update the invite link structure and explanation
The template link provided by the guide uses a different template than the one provided by discord's OAuth2 app page, so I modified it to use the one provided by discord. I also updated the URL part by part of the parsing by adding the explanation of the authorization part and I completely rewrote the link building part so that it explains hot to do it with OAuth2 page.