create a subsection in Threads for send via webhook
This addresses #914
The example is perhaps a little too long? I could see a valid argument made to simplify this by removing the requires, login, and listener; however it seems more apt to have in my mind.

I am also uncertain if my linking to the webhook page is done properly, do I link to the .md or .html file?