Clarification: "Client ID" into "Bot's Client ID"
In fact, it is not "your client id" when you refer to the reader, but rather "your bot's client id". 

For someone just starting off with discord bot programming, it may not be obvious that slash commands are, at the smallest scope, registered guild-wise, but never user-wise. Without this knowledge, it may feel justifiable to use particular *user's client ID*, instead of *bot's client ID*, that is actually needed. 

This addition clears the confusion.