Update cooldowns.md
Updated how the message to inform the user of the amount of time they need to wait before using the command again looks. Now it looks better with the format of the unix timestamp. Also added more clarity to add the code to the InteractionCreate event, not just anywhere in the main file.
@Joeyk710 - Helped with changes