Update index.js
Updated how the message to inform the user of the amount of time they need to wait before using the command again looks. Now it looks better with the format of the unix timestamp.
Also updated ${command.name} to ${command.data.name} so it will not return undefined. This request is just to update the resulting code with the guide.