Modifying example to search command subfolders
This portion of the guide is confusing because it does not work with the previous section title "Command Categories" which has the user move commands in to subcommand folders.

The changes I've made allow for the full commands folder to be searched for a matching command. This will work no matter the path to the file inside the command folder.